+++ PMFightAcademy.Admin/Controllers/ServicesController.cs
@@ -22,13 +22,15 @@ namespace PMFightAcademy.Admin.Controllers
     public class ServicesController : ControllerBase
     {
         private readonly IServiceService _serviceService;
+        private readonly IWorkWithIdService _checkId;
 
         /// <summary>
         /// Service Controller
         /// </summary>
-        public ServicesController(IServiceService serviceService)
+        public ServicesController(IServiceService serviceService,IWorkWithIdService checkId)
         {
             _serviceService = serviceService;
+            _checkId = checkId;
         }
 
         #region JS TILT
@@ -61,7 +63,7 @@ namespace PMFightAcademy.Admin.Controllers
         /// <see cref="HttpStatusCode.OK"/> add a coach to coaches
         /// <see cref="HttpStatusCode.NotFound"/> return lit of services</returns>
         /// <remarks> Use to Get all service, return services if  all is fine
-        /// NotFound if its is already registered
+        /// NotFound if its is service not  registered
         /// </remarks>
         /// <exception cref="NotImplementedException"></exception>
         [HttpGet]
@@ -84,18 +86,24 @@ namespace PMFightAcademy.Admin.Controllers
         /// <param name="serviceId"></param>
         /// <returns>
         /// <see cref="HttpStatusCode.OK"/> return service needed 
-        /// <see cref="HttpStatusCode.NotFound"/> if service not founded
+        /// <see cref="HttpStatusCode.NotFound"/>if service not founded
+        ///  <see cref="HttpStatusCode.BadRequest"/> if id is incorrect 
         /// </returns>
         /// <remarks> Use to Get Service, return service if  all is fine
-        /// NotFound if its is already registered
+        /// NotFound if its is  not  registered
         /// </remarks>
         /// <exception cref="NotImplementedException"></exception>
         [HttpGet("{serviceId}")]
         [ProducesResponseType(typeof(Service), (int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(string), (int)HttpStatusCode.NotFound)]
+        [ProducesResponseType(typeof(string), (int)HttpStatusCode.BadRequest)]
         public async Task<IActionResult> GetService(int serviceId)
         {
-            
+            if (!_checkId.IsCorrectId(serviceId))
+            {
+                return BadRequest("incorrect Id");
+            }
+
             var service = await _serviceService.TakeService(serviceId);
             if (service != null)
             { 
@@ -135,7 +143,7 @@ namespace PMFightAcademy.Admin.Controllers
         }
 
         /// <summary>
-        /// Add services 
+        /// Update services 
         /// </summary>
         /// <param name="service"></param>
         /// <param name="cancellationToken"></param>
@@ -151,6 +159,7 @@ namespace PMFightAcademy.Admin.Controllers
         [HttpPost("update")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(string), (int)HttpStatusCode.NotFound)]
+        [ProducesResponseType(typeof(string), (int)HttpStatusCode.BadRequest)]
         public async Task<IActionResult> UpdateService([FromBody] Service service, CancellationToken cancellationToken)
         {
             var update = await _serviceService.UpdateService(service, cancellationToken);
@@ -186,7 +195,7 @@ namespace PMFightAcademy.Admin.Controllers
         //}
 
         /// <summary>
-        /// Add services 
+        /// Delete services 
         /// </summary>
         /// <param name="serviceId"></param>
         /// <param name="cancellationToken"></param>
@@ -197,13 +206,19 @@ namespace PMFightAcademy.Admin.Controllers
         /// <remarks>
         /// Use to delete service
         /// Return ok if deleted and not found if BD have not this service
-        /// Conflict if its is already registered</remarks>
+        /// </remarks>
         /// <exception cref="NotImplementedException"></exception>
         [HttpDelete]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(string), (int)HttpStatusCode.NotFound)]
+        [ProducesResponseType(typeof(string), (int)HttpStatusCode.BadRequest)]
         public async Task<IActionResult> DeleteService(int serviceId, CancellationToken cancellationToken)
         {
+            if (!_checkId.IsCorrectId(serviceId))
+            {
+                return BadRequest("incorrect Id");
+            }
+
             var deleted = await _serviceService.DeleteService(serviceId, cancellationToken);
 
             if (deleted)
